fix bug in rpc_client, now can success connect and send rpc message!

+++ mRPC/include/net/timer_event.h
@@ -16,7 +16,7 @@ public:
 	using s_ptr = std::shared_ptr<TimerEvent>;
 
 public:
-	TimerEvent(int interval, bool is_repeated, const std::function<void()>& cb);
+	TimerEvent(int64_t interval, bool is_repeated, const std::function<void()>& cb);
 
 	int64_t getArriveTime() const { return m_arrive_time; }
