+++ mRPC/src/net/timer_event.cc
@@ -4,7 +4,7 @@
 
 MRPC_NAMESPACE_BEGIN
 
-TimerEvent::TimerEvent(int interval, bool is_repeated,
+TimerEvent::TimerEvent(int64_t interval, bool is_repeated,
                        const std::function<void()>& cb)
     : m_interval(interval)
     , m_is_repeated(is_repeated)
